scope-backend/pyproject.toml
[project]
name = "scope-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.23.2",
    "torch>=2.0.1",
    "transformers>=4.34.1",
    "scikit-learn",
    "pydantic>=2.4.2",
    "pydantic-settings>=2.0.3",
    "python-multipart>=0.0.6",
    "email-validator>=2.0.0",
    "sqlalchemy>=2.0.22",
    "alembic>=1.12.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "bcrypt>=4.0.0",
    "langchain>=0.0.335",
    "langchain-google-genai>=0.0.5",
    "langchain-community>=0.0.10",
    "faiss-cpu>=1.7.4",
    "pandas>=2.1.1",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
]
